requirements.txt
web3
termcolor
python-dotenv
requests
